refactor runner fixture into conftest.py
