+++ api/src/conftest.py
@@ -7,6 +7,12 @@ import logging
 
 from . import db, create_app
 
+from commands.people import create_account_cli
+from commands.app import create_app_cli
+from commands.courses import create_course_cli
+from commands.events import create_event_cli
+from commands.faker import create_faker_cli
+from commands.i18n import create_i18n_cli
 
 class AuthClient(FlaskClient):
     def __init__(self, *args, **kwargs):
@@ -47,3 +53,21 @@ def auth_client():
 @pytest.fixture
 def plain_client():
     yield from client_factory(FlaskClient)
+
+@pytest.fixture
+def runner():
+    app = create_app(os.getenv('CC_CONFIG') or 'test')
+    app.testing = True  # Make sure exceptions percolate out
+
+    db.drop_all()
+    db.create_all()
+
+    create_account_cli(app)
+    create_app_cli(app)
+    create_course_cli(app)
+    create_event_cli(app)
+    create_faker_cli(app)
+    create_i18n_cli(app)
+
+    yield app.test_cli_runner()
+
