+++ api/src/test_cli.py
@@ -3,37 +3,11 @@ import os
 import pytest
 import json
 import yaml
-from src import db, create_app
-from src.courses.models import Course, Diploma
-from src.i18n.models import Language, I18NValue, I18NLocale, I18NKey
-from src.people.models import Role
-from src.places.models import Country
-
-from commands.people import create_account_cli
-from commands.app import create_app_cli
-from commands.courses import create_course_cli
-from commands.events import create_event_cli
-from commands.faker import create_faker_cli
-from commands.i18n import create_i18n_cli
-
-
-@pytest.fixture
-def runner():
-    app = create_app(os.getenv('CC_CONFIG') or 'test')
-    app.testing = True  # Make sure exceptions percolate out
-
-    db.drop_all()
-    db.create_all()
-
-    create_account_cli(app)
-    create_app_cli(app)
-    create_course_cli(app)
-    create_event_cli(app)
-    create_faker_cli(app)
-    create_i18n_cli(app)
-
-    yield app.test_cli_runner()
-
+from . import db
+from .courses.models import Course, Diploma
+from .i18n.models import Language, I18NValue, I18NLocale, I18NKey
+from .people.models import Role
+from .places.models import Country
 
 def test_load_countries(runner):
     result = runner.invoke(args=['app', 'load-countries'])
